Poll HTCondor and filter out non-VM IPs
HTCondor keeps a VM in the status list even when it is shut down. If by
chance a new VM comes up with the same IP address, Elastiq might think
that it is idle since some time and kill it before it boots, because of
stale entries in condor_status. This patch should address the problem.